Выполнено продолжение рефакторинга

+++ src/main/java/ru/hits/attackdefenceplatform/public_interface/dashboard/TeamScoreChangeDto.java
@@ -6,8 +6,8 @@ import java.util.UUID;
 public record TeamScoreChangeDto(
         String teamName,
         Date submissionTime,
-        Integer pointsEarned,
-        Integer totalTeamPoints,
+        Double pointsEarned,
+        Double totalTeamPoints,
         String color
 ) {}
